Add files via upload
2. Use csv.Dictreader
Use the csv.DictReader method for the following:
a) Read the file and print a list of unique categories. Use the constant ‘Category’ in
your file reading. Hint: create a variable unique_list and put in the category value if it
is not already there.
b) Extend the above to print a list of the unique categories and the count of rows that
have that category.
c) Extend the above by making it a function that takes a file name and a header to be
more general.